chore(upgrade): change to ethers and updgrade dependencies
