+++ .husky/pre-push
@@ -1,6 +1,5 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
 
-npm run format
-npm run lint
-npx lint-staged
+npm run build
+npm run test
