+++ .prettierrc.json5
@@ -1,6 +1,6 @@
 {
-    trailingComma: 'none',
-    tabWidth: 4,
+    trailingComma: 'es5',
+    tabWidth: 2,
     semi: true,
     singleQuote: true,
     jsxSingleQuote: true,
