--- src/VerifierResponse.ts
@@ -1,3 +1,0 @@
-export default interface VerifierResponse {
-    trustedVerifiers: string[];
-}
