+++ src/VersionsManager.ts
@@ -1,27 +1,28 @@
 import semver from 'semver';
 
 export default class VersionsManager {
-    readonly componentVersion: string;
+  readonly componentVersion: string;
 
-    /**
-     * @param componentVersion - a semver of a component that uses the VersionsManager
-     */
-    constructor(componentVersion: string) {
-        if (semver.valid(componentVersion) == null) {
-            throw new Error('Component version is not valid');
-        }
-        this.componentVersion = componentVersion;
+  /**
+   * @param componentVersion - a semver of a component that uses the VersionsManager
+   */
+  constructor(componentVersion: string) {
+    if (semver.valid(componentVersion) == null) {
+      throw new Error('Component version is not valid');
     }
+    this.componentVersion = componentVersion;
+  }
 
-    /**
-     * @param version - the version of a dependency to compare against
-     * @return true if {@param version} is same or newer then {@link componentVersion}
-     */
-    isMinorSameOrNewer(version: string): boolean {
-        // prevent crash with some early verifiers (which are otherwise perfectly valid)
-        version = version.replace('_', '-');
+  /**
+   * @param version - the version of a dependency to compare against
+   * @return true if {@param version} is same or newer then {@link componentVersion}
+   */
+  isMinorSameOrNewer(version: string): boolean {
+    // prevent crash with some early verifiers (which are otherwise perfectly valid)
+    version = version.replace('_', '-');
 
-        const range = '^' + this.componentVersion;
-        return semver.satisfies(version, range);
-    }
+    const range = '^' + this.componentVersion;
+
+    return semver.satisfies(version, range);
+  }
 }
