--- src/dev/SendCallback.ts
@@ -1,6 +1,0 @@
-import { JsonRpcResponse } from 'web3-core-helpers';
-
-export type SendCallback = (
-    error: Error | null,
-    result?: JsonRpcResponse
-) => void;
