+++ src/index.ts
@@ -1,31 +1,9 @@
-import EnvelopingTransactionDetails from './types/EnvelopingTransactionDetails';
-import ContractInteractor from './ContractInteractor';
-import replaceErrors from './ErrorReplacerJSON';
-import PingResponse from './PingResponse';
-import TokenResponse from './TokenResponse';
-import VerifierResponse from './VerifierResponse';
+import EnvelopingTransactionDetails from '../types/EnvelopingTransactionDetails';
 import VersionsManager from './VersionsManager';
 
-export * from './dev/NetworkSimulatingProvider';
-export * from './dev/ProfilingProvider';
-export * from './dev/SendCallback';
-export * from './dev/WrapperProviderBase';
-export * from './types/EnvelopingConfig';
-export * from './types/EnvelopingTransactionDetails';
-export * from './types/RelayTransactionRequest';
-export * from './AmountRequired';
-export * from './Constants';
-export * from './ContractInteractor';
-export * from './Environments';
-export * from './Utils';
+export * from '../types/EnvelopingConfig';
+export * from '../types/EnvelopingTransactionDetails';
+export * from '../types/RelayTransactionRequest';
 export * from './VersionRegistry';
 
-export {
-    EnvelopingTransactionDetails,
-    ContractInteractor,
-    replaceErrors,
-    PingResponse,
-    TokenResponse,
-    VerifierResponse,
-    VersionsManager
-};
+export { EnvelopingTransactionDetails, VersionsManager };
