--- test/basic.test.ts
@@ -1,7 +1,0 @@
-import { assert } from 'chai';
-
-describe('Common test', () => {
-    it('Unit testing', () => {
-        assert.isTrue(true);
-    });
-});
