+++ tsconfig.json
@@ -1,43 +1,31 @@
 {
   "compilerOptions": {
-    "allowJs": true,
-    "resolveJsonModule": true,
-    "esModuleInterop": true,
-    "target": "es2017",
+    "lib": ["es2020", "esnext"],
+    "target": "es2020",
     "module": "commonjs",
     "strict": true,
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "outDir": "dist",
+    "declaration": true,
+    "baseUrl": "./",
+    "allowJs": true,
     "strictNullChecks": false,
     "strictPropertyInitialization": false,
     "moduleResolution": "node",
     "alwaysStrict": true,
-    "declaration": true,
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
     "forceConsistentCasingInFileNames": true,
     "skipLibCheck": true,
-    "outDir": "dist",
-    "typeRoots": [
-      "node_modules/@types"
-    ],
+    // "typeRoots": [
+    //   "node_modules/@types"
+    // ],
     "types": [
       "node",
       "mocha",
       "chai"
     ],
-    "lib": [
-      "es2018",
-      "esnext"
-    ],
     "sourceMap": true
-  },
-  "include": [
-    "**/*.ts",
-    "src/**/*"
-  ],
-  "exclude": [
-    "dist",
-    "node_modules",
-    "build",
-    "test"
-  ]
+  }
 }
